updated dependancy, removed unecessary reference,

+++ MonsterCast/SQLiteAsync.cs
@@ -107,7 +107,7 @@ namespace MonsterCast
 				var conn = GetConnection ();
 				using (conn.Lock ()) {
 					foreach (Type type in types) {
-						int aResult = conn.CreateTable (type, createFlags);
+						int aResult = (int)conn.CreateTable (type, createFlags);
 						result.Results[type] = aResult;
 					}
 				}
